+++ src/test/java/org/opensearch/plugin/insights/core/service/categorizer/SearchQueryCategorizerTests.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.plugin.insights.core.service.categorizor;
+package org.opensearch.plugin.insights.core.service.categorizer;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -15,6 +15,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.opensearch.plugin.insights.QueryInsightsTestUtils.generateQueryInsightRecords;
+import static org.opensearch.plugin.insights.core.service.categorizer.SearchQueryAggregationCategorizer.AGGREGATION_TYPE_TAG;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -114,7 +115,7 @@ public final class SearchQueryCategorizerTests extends OpenSearchTestCase {
         verify(searchQueryCategorizer.getSearchQueryCounters().getAggCounter()).add(valueCaptor.capture(), tagsCaptor.capture());
 
         double actualValue = valueCaptor.getValue();
-        String actualTag = (String) tagsCaptor.getValue().getTagsMap().get("agg_type");
+        String actualTag = (String) tagsCaptor.getValue().getTagsMap().get(AGGREGATION_TYPE_TAG);
 
         assertEquals(1.0d, actualValue, 0.0001);
         assertEquals(MULTI_TERMS_AGGREGATION, actualTag);
